Refactor `core` module, to break `core_state` into a new source file.
